Divide Character, BattleArena and Weapon into 3 different classes
